refactor(desktop): extract focus debug logic
